+++ app_template/settings.py
@@ -1,4 +1,7 @@
 class Settings:
+
+    # APP DB Settings.
+    # Required if you add any DB Mapping in conf/resource_mapping.yaml
     DATABASE = {
         "engine": "",  # Allowed values are "postgres", "mysql", "sqlite"
         "name": "",
@@ -6,18 +9,6 @@ class Settings:
         "password": "",
         "host": "",
         "port": ""
-    }
-
-    # These APIs would not be hit during load test
-    # Strings are matched via Regex Mechanism
-    EXCLUDE_URLS = ["/logout", "/login"]
-
-    SERVER_URL = {
-        "protocol": "",       # If left empty, would be picked from swagger. If not in swagger, would be "http"
-        "host": "",           # if left empty, would be picked from swagger. If not there, would be localhost
-
-        # In swagger, we search for info/url and basePath  for this setting
-        "api_url": "",       # if left empty, would be picked from swagger. If not there, would be blank string.
     }
 
     # Page Query Parameters
@@ -32,11 +23,7 @@ class Settings:
     # Only hit APIs which match Tags
     ONLY_TAG_API = False
 
-    # Custom Ordering Dependency
-    # This is an array of 2-pair tuples, with second operation being dependent on first operation
-    # These are operation IDs --> Eg: [ (petCreate, petList), (petCreate, petRetrieve) ]
-    SWAGGER_OPERATION_DEPENDENCIES = []
-
+    # INFLUX DB Setting. Required.
     INFLUX = {
         "database": "",
         "host": "",