+++ atlas/modules/constants.py
@@ -53,6 +53,7 @@ STRING = "string"
 BOOLEAN = "boolean"
 ARRAY = "array"
 ENUM = "enum"   # Strictly not a type, but is associated with both strings and numbers
+FILE = "file"
 QUERY_TYPES = {INTEGER, NUMBER, STRING, BOOLEAN, ARRAY}    # Valid types for query Parameters
 
 # Integer Options
@@ -76,6 +77,7 @@ UUID = "uuid"
 URI = "uri"
 URL = "url"
 SLUG = "slug"
+STRING_JSON = "json"
 
 # Array Options
 MIN_ITEMS = "minItems"