+++ atlas/modules/resource_creator/creators.py
@@ -157,6 +157,11 @@ class AutoGenerator(mixins.YAMLReadWriteMixin):
             self.parse_reference_properties(ref_name, schema.get(swagger_constants.PROPERTIES))
 
     def get_ref_name_and_config(self, ref):
+
+        if not isinstance(ref, str):
+            print(f"\nWARNING: Only string references supported. Found: {ref}\n")
+            return
+
         ref_config = utils.resolve_reference(self.specs, ref)
         ref_name = ref.split("/")[-1]
         self.parse_reference(ref_name, ref_config)
@@ -164,7 +169,7 @@ class AutoGenerator(mixins.YAMLReadWriteMixin):
     def parse_reference(self, ref_name, ref_config):
 
         if ref_name in self.processed_refs:
-            return      # This has already been processed, so need to do it again
+            return      # This has already been processed, so no need to do it again
 
         for element in ref_config.get(swagger_constants.ALL_OF, []):
             self.resolve_all_of_element(ref_name, element)
@@ -172,6 +177,9 @@ class AutoGenerator(mixins.YAMLReadWriteMixin):
         self.parse_reference_properties(ref_name, ref_config.get(swagger_constants.PROPERTIES, {}))
 
     def parse_reference_properties(self, ref_name, properties):
+
+        # By adding it to processed list before even processing, we avoid cycles
+        self.processed_refs.add(ref_name)
 
         for key, value in properties.items():
             resource = ""
@@ -189,8 +197,6 @@ class AutoGenerator(mixins.YAMLReadWriteMixin):
                 resource_alias = self.resource_map_resolver.get_alias(resource)
                 value[swagger_constants.RESOURCE] = resource_alias
                 self.resource_params.add(resource_alias)
-
-        self.processed_refs.add(ref_name)
 
     def parse(self):
 
@@ -214,7 +220,7 @@ class AutoGenerator(mixins.YAMLReadWriteMixin):
                     # Operation IDs are used as primary key throughout application
                     op_id = method_config.get(swagger_constants.OPERATION)
                     if not op_id:
-                        method_config[swagger_constants.OPERATION] = self.operation_id_name(url, method)
+                        method_config[swagger_constants.OPERATION] = utils.operation_id_name(url, method)
 
                     if parameters:
                         self.parse_params(parameters, url)
@@ -227,32 +233,6 @@ class AutoGenerator(mixins.YAMLReadWriteMixin):
 
         for ref_name, ref_config in self.specs.get(swagger_constants.DEFINITIONS, {}).items():
             self.parse_reference(ref_name, ref_config)
-
-    @staticmethod
-    def operation_id_name(url, method) -> str:
-        """
-        Generate the name for Operation ID
-
-        Logic:
-            user/       - (user_create, user_list)
-            user/{id}   - (user_read, user_update, user_delete)
-            user/{id}/action - (user_action with above logic)
-        """
-
-        url_fragments = [fragment for fragment in url.split("/") if fragment]
-
-        op_name_array = [
-            url_element for url_element in url_fragments if not url_element.startswith("{")
-        ]
-
-        if method == swagger_constants.DELETE:
-            op_name_array.append("delete")
-        elif url_fragments[-1].startswith("{"):
-            op_name_array.append("read" if method == swagger_constants.GET else "update")
-        else:
-            op_name_array.append("list" if method == swagger_constants.GET else "create")
-
-        return "_".join(op_name_array)
 
     def update(self):