+++ Changelog.md
@@ -2,6 +2,35 @@ Changelog
 ========
 
 
+8.0.0
+-----
+
+One of the biggest releases up to date.
+It is recommended you create a new project with this release and copy your old settings here.
+
+*Breaking Changes*
+- Settings have been refactored.
+Several settings have been renamed, and some of the settings now reside in `conf/conf.py`.
+- Several Setting Inputs have been changed
+- Swagger Operation ID naming schema has been changed
+
+*Features*
+- Added option for custom scenarios
+- All Swagger Operation inputs are now consolidated in one format
+- Can generate OP_KEY for all Swagger routes by using `python manage.py generate_routes`
+- Several Swagger Keys can now be used as field names in References
+- Added Dummy JSON Provider (type: string, format: json --> returns {})
+- Added Regex Pattern support in Providers for string
+- Add support for adding File Sub types from Hooks
+- Comprehensive documentation updates
+
+*Bug fixes*
+- Fix `Enum` options in provider if their choice was false-y value
+- String and URL Providers enhanced to correctly support MAX characters
+- File Provider fixed
+- Resource Values were deleted even if their API was failing. This has been fixed
+
+
 7.0.0
 -----